requirements.txt
appdirs==1.4.4
attrs==23.2.0
black==24.2.0
blosc2==2.5.1
botocore==1.34.54
certifi==2022.12.7
chamfer-distance @ git+https://github.com/otaheri/chamfer_distance@f86f6f7cadd3aca642704573d1626c67ca2e2846
charset-normalizer==2.1.1
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.7.0
filelock==3.9.0
flake8==7.0.0
flake8-bugbear==24.2.6
flake8-comprehensions==3.14.0
fonttools==4.49.0
fsspec==2024.2.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.42
h5py==3.10.0
huggingface-hub==0.21.3
idna==3.4
imageio==2.34.0
imath==0.0.2
iopath==0.1.10
isort==4.3.21
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.3
mccabe==0.7.0
mpmath==1.3.0
msgpack==1.0.8
mypy-extensions==1.0.0
ndindex==1.8
networkx==3.2.1
ninja==1.11.1.1
numexpr==2.9.0
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
opencv-python==4.9.0.80
OpenEXR==3.2.3
packaging==23.2
pandas==2.2.1
pathspec==0.12.1
pillow==10.2.0
Pillow-SIMD==9.0.0.post1
platformdirs==4.2.0
portalocker==2.8.2
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pycodestyle==2.11.1
pyflakes==3.2.0
pykitti==0.3.1
pyparsing==3.1.1
python-dateutil==2.9.0.post0
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@2f11ddc5ee7d6bd56f2fb6744a16776fab6536f7
pytz==2024.1
PyYAML==6.0.1
requests==2.28.1
safetensors==0.4.2
scipy==1.12.0
sentry-sdk==1.40.6
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
tables==3.9.2
tabulate==0.9.0
termcolor==2.4.0
timm==0.9.16
torch==2.2.0+cu118
torchaudio==2.2.0+cu118
torchvision==0.17.0+cu118
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.8.0
tzdata==2024.1
urllib3==1.26.13
wandb==0.16.3
xformers==0.0.24+cu118
yacs==0.1.8
